Create some boilerplate classes for the data structure
I created a few empty classes in order to lay out how the data structure will be formatted and programmed.
I am currently working on starting the data structure.
I need to finish creating the boilerplate classes for the data structure and I still need to implement the actual functionality of the structure.