Started adding functionality to the data structure
What I did: I restructured the code files to make it more readable. I started adding functionality to the data structure.
What I still need to do: I have found a pretty big flaw in the planned structure of the data structure, but have found a solution to it and already started implementing it.