Added new testing system to benchmark the data structure
I made a new testing system to allow me to more easily benchmark and time the performance of the data structure.

Now I need to add a way for me to take the water simulation and add it to a specific grid level and have its data accessible to other grid levels. I also need to add soil water.